requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PySide6~=6.9.0
pytz~=2025.2
SQLAlchemy~=2.0.41
python-dateutil~=2.9.0.post0
colorlog~=6.9.0
requests~=2.32.3